:art: Apply pr comments
